requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
beautifultable==1.1.0
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
dash==2.16.1
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.18.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.2
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
hugchat==0.4.7
huggingface-hub==0.21.4
idna==2.8
importlib_metadata==7.0.2
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
loguru==0.7.2
macos-notifications==0.2.0
MarkupSafe==2.1.5
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
nest-asyncio==1.6.0
numpy==1.26.4
openai==0.28.0
packaging==23.2
pandas==2.2.1
plotly==5.19.0
preshed==3.0.9
pyarrow==15.0.2
pyarrow-hotfix==0.6
pydantic==2.7.1
pydantic_core==2.18.2
pydbpedia==0.0.3
pyparsing==3.1.2
pysondb==1.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rdflib==7.0.0
requests==2.22.0
requests-toolbelt==1.0.0
retrying==1.3.4
scikit-learn==1.4.1.post1
scipy==1.12.0
shortuuid==1.0.13
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-dbpedia-spotlight==0.2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
tenacity==8.2.3
thinc==8.2.3
threadpoolctl==3.3.0
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.25.11
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
